fix: separated strategy description component to be shared
