+++ packages/bot-web-ui/src/components/dashboard/tutorial-tab/quick-strategy-content/index.scss
@@ -53,6 +53,7 @@
     }
     &__breadcrumb {
         display: flex;
+        margin-bottom: 1.6rem;
 
         &__clickable {
             cursor: pointer;
