+++ packages/bot-web-ui/src/components/quick-strategy/form-wrappers/desktop-form-wrapper.tsx
@@ -9,7 +9,7 @@ import { localize } from '@deriv/translations';
 import { useDBotStore } from 'Stores/useDBotStore';
 import { STRATEGIES } from '../config';
 import FormTabs from './form-tabs';
-import StrategyDescription from './strategy-description';
+import StrategyTabContent from './strategy-tab-content';
 import useQsSubmitHandler from './useQsSubmitHandler';
 import '../quick-strategy.scss';
 
@@ -104,7 +104,7 @@ const FormWrapper: React.FC<TDesktopFormWrapper> = observer(({ children }) => {
                             onChange={handleTabChange}
                             description={strategy?.long_description}
                         />
-                        <StrategyDescription formfields={children} active_tab={activeTab} />
+                        <StrategyTabContent formfields={children} active_tab={activeTab} />
                     </ThemedScrollbars>
                     {activeTab === 'TRADE_PARAMETERS' && (
                         <div className='qs__body__content__footer'>
