+++ packages/bot-web-ui/src/components/quick-strategy/types.ts
@@ -48,7 +48,7 @@ export type TDescriptionItem = {
 export type TStrategy = {
     name: string;
     label: string;
-    description: string;
+    description?: string;
     long_description?: TDescriptionItem[];
     fields: TConfigItem[][];
 };
